Having the unrecognized input card correctly included
