--- Source/Icebreaker/Helpers/AdaptiveCards/WelcomeNewMemberAdaptiveCard.json
@@ -1,65 +1,0 @@
-﻿{
-  "type": "AdaptiveCard",
-  "body": [
-    {
-      "type": "Container",
-      "items": [
-        {
-          "type": "TextBlock",
-          "size": "Medium",
-          "weight": "Bolder",
-          "text": "%salutationText%"
-        }
-      ]
-    },
-    {
-      "type": "Container",
-      "items": [
-        {
-          "type": "Image",
-          "horizontalAlignment": "Center",
-          "url": "%welcomeCardImageUrl%",
-          "width": "424px",
-          "height": "248px"
-        },
-        {
-          "type": "TextBlock",
-          "text": "%introMessagePart1%",
-          "wrap": true
-        },
-        {
-          "type": "TextBlock",
-          "text": "%introMessagePart2%",
-          "wrap": true,
-          "spacing": "Medium"
-        },
-        {
-          "type": "TextBlock",
-          "text": "%introMessagePart3%",
-          "wrap": true, 
-          "spacing": "Medium"
-        }
-      ]
-    }
-  ],
-  "actions": [
-    {
-      "type": "Action.OpenUrl",
-      "title": "%tourButtonText%",
-      "url": "%tourUrl%"
-    },
-    {
-      "type": "Action.Submit",
-      "title": "%pauseMatchesText%",
-      "data": {
-        "msteams": {
-          "type": "messageBack",
-          "displayText": "%pauseMatchesText%",
-          "text": "optout"
-        }
-      }
-    }
-  ],
-  "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
-  "version": "1.0"
-}
